requirements.txt
mysql-connector-python>=8.0
spacy>=3.7
fastapi>=0.110
uvicorn[standard]>=0.24
